backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Warehouse management backend with real-time log processing"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.3.0",
    "kafka-python>=2.2.15",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "pytest>=7.4.0",
    "pytest-django>=4.8.0",
    "pytest-factoryboy>=2.6.0",
    "factory-boy>=3.3.0",
    "freezegun>=1.4.0",
    "pytest-cov>=4.1.0",
]
